# repositories/projects/69c5b89a-13f6-4821-8b8d-add4bd1c0724/ShipTown_e5fd0c33-9c52-4a81-9fb0-df619eb471d0/public/docs/openapi.yaml
openapi: 3.0.3
info:
  title: ShipTown
  description: 'API documentation for ShipTown'
  version: 1.0.0
servers:
  -
    url: 'https://demo.myshiptown.com'
paths:
...
  /api/modules/api2cart/connections:
    get:
      summary: ''
      operationId: getApiModulesApi2cartConnections
      description: ''
      parameters: []
      responses:
        401:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  message: Unauthenticated.
                properties:
                  message:
                    type: string
                    example: Unauthenticated.
      tags:
        - API
    post:
      summary: ''
      operationId: postApiModulesApi2cartConnections
      description: ''
      parameters: []
      responses: {  }
      tags:
        - API
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                location_id:
                  type: string
                  description: ''
                  example: at
                url:
                  type: string
                  description: 'Must be a valid URL.'
                  example: 'http://www.mertz.org/'
                type:
                  type: string
                  description: ''
                  example: enim
                bridge_api_key:
                  type: string
                  description: ''
                  example: iure
              required:
                - location_id
                - url
                - type
                - bridge_api_key
  '/api/modules/api2cart/connections/{id}':
    delete:
      summary: ''
      operationId: deleteApiModulesApi2cartConnectionsId
      description: ''
      parameters: []
      responses: {  }
      tags:
        - API
    parameters:
      -
        in: path
        name: id
        description: 'The ID of the connection.'
        example: 20
        required: true
        schema:
          type: integer
...
  /api/modules/dpd-ireland/connections:
    get:
      summary: ''
      operationId: getApiModulesDpdIrelandConnections
      description: ''
      parameters: []
      responses:
        401:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  message: Unauthenticated.
                properties:
                  message:
                    type: string
                    example: Unauthenticated.
      tags:
        - API
    post:
      summary: ''
      operationId: postApiModulesDpdIrelandConnections
      description: ''
      parameters: []
      responses: {  }
      tags:
        - API
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                user:
                  type: string
                  description: ''
                  example: nihil
                token:
                  type: string
                  description: ''
                  example: non
                password:
                  type: string
                  description: ''
                  example: '"Q|ilQ|vw'
                live:
                  type: boolean
                  description: ''
                  example: false
                contact:
                  type: string
                  description: ''
                  example: qui
                contact_telephone:
                  type: string
                  description: ''
                  example: amet
                contact_email:
                  type: string
                  description: ''
                  example: legros.carmela@example.com
                business_name:
                  type: string
                  description: ''
                  example: cum
                address_line_1:
                  type: string
                  description: ''
                  example: tenetur
                address_line_2:
                  type: string
                  description: ''
                  example: nulla
                address_line_3:
                  type: string
                  description: ''
                  example: quae
                address_line_4:
                  type: string
                  description: ''
                  example: nisi
                country_code:
                  type: string
                  description: ''
                  example: IE
                  enum:
                    - IE
                    - IRL
                    - UK
                    - GB
              required:
                - user
                - token
                - password
                - live
                - contact
                - contact_telephone
                - address_line_3
                - address_line_4
                - country_code
  '/api/modules/dpd-ireland/connections/{id}':
    delete:
      summary: ''
      operationId: deleteApiModulesDpdIrelandConnectionsId
      description: ''
      parameters: []
      responses: {  }
      tags:
        - API
    parameters:
      -
        in: path
        name: id
        description: 'The ID of the connection.'
        example: sit
        required: true
        schema:
          type: string
...
  /api/modules/magento-api/connections:
    get:
      summary: ''
      operationId: getApiModulesMagentoApiConnections
      description: ''
      parameters: []
      responses:
        401:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  message: Unauthenticated.
                properties:
                  message:
                    type: string
                    example: Unauthenticated.
      tags:
        - API
    post:
      summary: ''
      operationId: postApiModulesMagentoApiConnections
      description: ''
      parameters: []
      responses: {  }
      tags:
        - API
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                base_url:
                  type: string
                  description: 'Must be a valid URL.'
                  example: 'http://kulas.net/sequi-nobis-quis-cum-magni-non-ipsa-reiciendis'
                magento_store_id:
                  type: number
                  description: ''
                  example: 521588.4303323
                tag:
                  type: string
                  description: ''
                  example: null
                pricing_source_warehouse_id:
                  type: string
                  description: ''
                  example: null
                api_access_token:
                  type: string
                  description: ''
                  example: vel
              required:
                - base_url
                - magento_store_id
                - api_access_token
  '/api/modules/magento-api/connections/{id}':
    put:
      summary: ''
      operationId: putApiModulesMagentoApiConnectionsId
      description: ''
      parameters: []
      responses: {  }
      tags:
        - API
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                base_url:
                  type: string
                  description: 'Must be a valid URL.'
                  example: 'http://www.zulauf.biz/aliquam-nam-sunt-corporis-quod-commodi.html'
                magento_store_id:
                  type: number
                  description: ''
                  example: 165102588.14207
                tag:
                  type: string
                  description: ''
                  example: a
                pricing_source_warehouse_id:
                  type: string
                  description: ''
                  example: totam
                api_access_token:
                  type: string
                  description: ''
                  example: quo
              required:
                - base_url
                - magento_store_id
                - tag
                - pricing_source_warehouse_id
                - api_access_token
    delete:
      summary: ''
      operationId: deleteApiModulesMagentoApiConnectionsId
      description: ''
      parameters: []
      responses: {  }
      tags:
        - API
    parameters:
      -
        in: path
        name: id
        description: 'The ID of the connection.'
        example: 13
        required: true
        schema:
          type: integer
...
  '/api/modules/magento2msi/connections/{id}':
    put:
      summary: ''
      operationId: putApiModulesMagento2msiConnectionsId
      description: ''
      parameters: []
      responses: {  }
      tags:
        - API
    delete:
      summary: ''
      operationId: deleteApiModulesMagento2msiConnectionsId
      description: ''
      parameters: []
      responses: {  }
      tags:
        - API
    parameters:
      -
        in: path
        name: id
        description: 'The ID of the connection.'
        example: omnis
        required: true
        schema:
          type: string
...
  /api/modules/rms_api/connections:
    get:
      summary: ''
      operationId: getApiModulesRms_apiConnections
      description: ''
      parameters: []
      responses:
        401:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  message: Unauthenticated.
                properties:
                  message:
                    type: string
                    example: Unauthenticated.
      tags:
        - API
    post:
      summary: ''
      operationId: postApiModulesRms_apiConnections
      description: ''
      parameters: []
      responses: {  }
      tags:
        - API
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                location_id:
                  type: string
                  description: 'This field is required when <code>id</code> is <code>null</code>.'
                  example: null
                url:
                  type: string
                  description: 'This field is required when <code>id</code> is <code>null</code>.  Must be a valid URL.'
                  example: 'http://www.ryan.com/'
                username:
                  type: string
                  description: 'This field is required when <code>id</code> is <code>null</code>.'
                  example: null
                password:
                  type: string
                  description: 'This field is required when <code>id</code> is <code>null</code>.'
                  example: null
                warehouse_code:
                  type: string
                  description: 'Must not be greater than 5 characters.'
                  example: jft
  '/api/modules/rms_api/connections/{id}':
    delete:
      summary: ''
      operationId: deleteApiModulesRms_apiConnectionsId
      description: ''
      parameters: []
      responses: {  }
      tags:
        - API
    parameters:
      -
        in: path
        name: id
        description: 'The ID of the connection.'
        example: 20
        required: true
        schema:
          type: integer
...
components:
  securitySchemes:
    default:
      type: http
      scheme: bearer
      description: 'You can retrieve your token by visiting <a href="/settings/api" target="_blank"><strong>Menu > Setting > API > Create API Token</strong></a>.'
security:
  -
    default: []